Fix long types
new version 0.02

+++ StepControl.h
@@ -6,20 +6,20 @@
  	public:
  		StepControl(int en, int step, int dir, int feth);
  		void setSpeeds(int speeds);
-	 	void moveTo(int degrees);
-	 	void moveToDistance(int mm);
+	 	void moveTo(long long degrees);
+	 	void moveToDistance(double mm);
 	 	void go(boolean g);
  	private:
  		boolean _g;
- 		int _mm;
+ 		double _mm;
  		int _feth;
  		int _enpin;
  		int _steppin;
  		int _dirpin;
- 		int _degrees;
+ 		long long _degrees;
  		int _speeds;
  		int microsstep;
- 		int i;
+ 		long long i;
 
  };
 #endif
